+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - integration with DataSync for ingestion of large volumes of data
-- when using identityStoreId, identityStoreRegion is also required
+
+## [1.0.7] - 2024-01-15
+
+### Fixed
+- when using identityStoreId, identityStoreRegion and identityStoreAccountId are also required
+- split refresh_token to a new cookie
 
 ## [1.0.6] - 2023-12-15